requirements.txt
bcrypt==3.1.4
cffi==1.11.2
Django==1.10
pycparser==2.18
python-dateutil==2.6.1
six==1.11.0
